Adds a server which writes and reads saved codes to a given MongoDB database.
Modifies the client to write/read saved codes to the server rather than through local storage.

Copies the client throttling logic from the RaspPi to the repo.

Comments out the THREE.js code on the client. This was done on the RaspPi.

Modifies .gitignore so the database path/username/password are not source controlled.